Gearman migration, moving to docker compose to start
Moving gearman workers to run in containers and register itself with the core container.

Moving to using docker compose to build and run the containers.

Using a shared volume to share files between core and workers.

Remove logs and archive directories that are part of the deployment but not the source.

Other cleanup of code.